requirements.txt
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.14.0
streamlit>=1.31.0
scikit-learn>=1.3.0
tensorflow>=2.15.0
keras>=2.15.0
transformers>=4.36.0
torch>=2.1.0
requests>=2.31.0
python-dotenv>=1.0.0
PyYAML>=6.0
openpyxl>=3.1.0
pandas-ta>=0.3.14b
newsapi-python>=0.2.7
twilio>=8.10.0
scipy>=1.11.0
ta>=0.11.0
